+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+1.1.0, 2026-02-09:
+- Payment by redirection compatibility with Hyva Checkout.
+- Bug fix: Fix embedded payment form behavior when invalid amount.
+
 1.0.1, 2025-12-18:
 - Bug fix: Do not process payment if order was not created.
 - Bug fix: Fix payment form token refresh when placing order.